add cocos attr to the script element in templates
To build project.json path, CCBoot.js is looking for a script element
with a cocos attribute. If element is not found, it loads project.json
from the current URL.
cocos attribute should be added in the template, because it is the
default CCBoot.js' behaviour, and usually, project.json is located with
game.min.js file.
https://github.com/cocos2d/cocos2d-html5/blob/develop/CCBoot.js#L2159